refact : reorganize project structure based on MVC design pattern
