+++ mutism/app/src/main/java/com/example/mutism/controller/noiseSelectPage/NoiseViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.mutism
+package com.example.mutism.controller.noiseSelectPage
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
